Create Magic Squares Program
"A magic square is a square matrix where all rows, columns, and diagonals sum to the same value. For example, the Dürer etching shown above includes the following 4 x 4 magic square with sum 34."
Wrote a program that uses a 2-D array to implement the Siamese method for constructing magic squares.